refactor: share warm runner health helpers
